requirements.txt
python3-resolver == 1.0.3
dnspython == 2.7.0
urllib3 == 2.5.0
